Implement Caffeine Coach, exact fractional servings, and visual identity refresh (v2.3.0)
- **Caffeine Coach**:
    - Implement a heuristic alertness model inspired by sleep research (homeostatic pressure, circadian rhythm, and caffeine pharmacodynamics).
    - Add `CaffeineCoach` utility to suggest "minimum effective, optimally-timed" doses.
    - Integrate coach recommendations into the home screen via `CaffeineCoachChip` and a detailed `CaffeineCoachSheet`.
    - Support personalized coaching by consuming Health Connect sleep data (actual sleep duration and wake times).

- **Logging & Catalog**:
    - Support exact fractional serving amounts (e.g., 1.5 cups) by migrating `ConsumptionEntry.quantity` and database schema from `Int` to `Double`.
    - Add `ServingQuantityStepper` with text-input support for precise figures and "quick pick" shortcuts for common volumes/weights.
    - Implement `DrinkCatalogSync` to merge new assets into existing databases (adds Moka Pot coffee and gram units for chocolate).

- **Design & UX**:
    - Introduce "Pulse Cup" branding: new adaptive app icon, Material You themed-icon support, and animated splash screen (Android 12+).
    - Refine typography by adding Montserrat-based expressive emphasized roles to the Material3 scale.
    - Add `hc_dismissed_records` tombstone system to prevent deleted Health Connect records from reappearing on subsequent imports.
    - Smooth UI transitions, including an expressive splash exit animation and skeleton-to-content fading for entry details.

- **Technical Details**:
    - Upgrade `versionCode` to 10 and `versionName` to `2.3.0`.
    - Migrate database to version 12 to support fractional quantities and dismissed HC records.
    - Optimize Health Connect writes using batched inserts.
    - Add comprehensive unit tests for `CaffeineCoach` and `WeeklySleepRota`.

+++ app/src/main/java/com/uc/caffeine/data/dao/ConsumptionLogDao.kt
@@ -29,7 +29,7 @@ interface ConsumptionLogDao {
     suspend fun updateEntryById(
         entryId: Int,
         caffeineMg: Int,
-        quantity: Int,
+        quantity: Double,
         unitKey: String,
         unitCaffeineMg: Double,
         startedAtMillis: Long,
